Refactor: #201 - captureTextView 함수 리턴형 수정
