+++ 2024-MacC-M4-6princess/2024-MacC-M4-6princess.xcodeproj/xcuserdata/piri.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -130,8 +130,8 @@
             filePath = "2024-MacC-M4-6princess/Source/View/DesignFrame/DFModifyView.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "169"
-            endingLineNumber = "169"
+            startingLineNumber = "176"
+            endingLineNumber = "176"
             landmarkName = "imageView"
             landmarkType = "24">
          </BreakpointContent>
@@ -162,8 +162,8 @@
             filePath = "2024-MacC-M4-6princess/Source/View/DesignFrame/DFModifyView.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "174"
-            endingLineNumber = "174"
+            startingLineNumber = "181"
+            endingLineNumber = "181"
             landmarkName = "imageView"
             landmarkType = "24">
          </BreakpointContent>
@@ -178,8 +178,8 @@
             filePath = "2024-MacC-M4-6princess/Source/View/DesignFrame/DFModifyView.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "180"
-            endingLineNumber = "180"
+            startingLineNumber = "187"
+            endingLineNumber = "187"
             landmarkName = "imageView"
             landmarkType = "24">
          </BreakpointContent>
@@ -194,8 +194,8 @@
             filePath = "2024-MacC-M4-6princess/Source/View/DesignFrame/DFModifyView.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "194"
-            endingLineNumber = "194"
+            startingLineNumber = "201"
+            endingLineNumber = "201"
             landmarkName = "imageView"
             landmarkType = "24">
          </BreakpointContent>
@@ -230,6 +230,22 @@
             endingLineNumber = "22"
             landmarkName = "makeUIViewController(context:)"
             landmarkType = "7">
+         </BreakpointContent>
+      </BreakpointProxy>
+      <BreakpointProxy
+         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
+         <BreakpointContent
+            uuid = "0519FBD7-04EC-4552-A8B1-88DBFAE736DA"
+            shouldBeEnabled = "No"
+            ignoreCount = "0"
+            continueAfterRunningActions = "No"
+            filePath = "2024-MacC-M4-6princess/Source/View/DesignFrame/DFModifyView.swift"
+            startingColumnNumber = "9223372036854775807"
+            endingColumnNumber = "9223372036854775807"
+            startingLineNumber = "79"
+            endingLineNumber = "79"
+            landmarkName = "body"
+            landmarkType = "24">
          </BreakpointContent>
       </BreakpointProxy>
    </Breakpoints>
